Further changes regarding early parsing
